安监版本开发——api接入开发

+++ src/main/java/com/xingkong/lyn/controller/anjian/UserController.java
@@ -4,14 +4,19 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 import com.xingkong.lyn.common.AjaxResults;
 import com.xingkong.lyn.common.constant.PasswordConstant;
+import com.xingkong.lyn.common.constant.ReviewConstant;
+import com.xingkong.lyn.entity.anjian.Review;
 import com.xingkong.lyn.entity.anjian.User;
 import com.xingkong.lyn.model.UserInfo;
 import com.xingkong.lyn.service.IUserInfo;
+import com.xingkong.lyn.service.anjian.IReview;
 import com.xingkong.lyn.service.anjian.IUser;
 import com.xingkong.lyn.util.EncodeUtil;
 import com.xingkong.lyn.util.OtherUtil;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
+import org.apache.shiro.subject.Subject;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -22,6 +27,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -35,6 +41,9 @@ public class UserController {
     @Resource
     private IUser userService;
 
+    @Resource
+    private IReview reviewService;
+
     @RequestMapping("/web/manage/user/add")
 //    @RequiresPermissions("userinfo:add")
     public Object add(@RequestBody User user){
@@ -44,7 +53,12 @@ public class UserController {
             ajaxResults.setMsg("参数错误");
         }
         if (null == userService.findByName(user.getUserName())) {
-            userService.addUser(user);
+            if (commitReview(user, ReviewConstant.OPERATE_ADD)) {
+                ajaxResults.setCode(0);
+                ajaxResults.setMsg("操作已提交，等待审核");
+            } else {
+                userService.addUser(user);
+            }
         } else {
             ajaxResults.setCode(0);
             ajaxResults.setMsg("用户已存在");
@@ -73,7 +87,12 @@ public class UserController {
         AjaxResults ajaxResults = new AjaxResults();
         Long[] arr = StringUtils.isBlank(id)? null: OtherUtil.parseStringtoLong(id);
         List<Long> ids = Arrays.asList(arr);
-        userService.deleteList(ids);
+        if (commitReviews(ids, ReviewConstant.OPERATE_DELETE)) {
+            ajaxResults.setCode(0);
+            ajaxResults.setMsg("操作已提交，等待审核");
+        } else {
+            userService.deleteList(ids);
+        }
         return ajaxResults;
     }
 
@@ -85,7 +104,12 @@ public class UserController {
             ajaxResults.setCode(0);
             ajaxResults.setMsg("参数错误");
         }
-        userService.updateUser(user);
+        if (commitReview(user, ReviewConstant.OPERATE_MODIFY)) {
+            ajaxResults.setCode(0);
+            ajaxResults.setMsg("操作已提交，等待审核");
+        } else {
+            userService.updateUser(user);
+        }
         return ajaxResults;
     }
 
@@ -93,5 +117,44 @@ public class UserController {
 //    @RequiresPermissions("userinfo:detail")
     public Object detail(Long id){
         return userInfoService.findById(id);
+    }
+
+    private boolean commitReview(User user, String operate) {
+        Subject currentUser = SecurityUtils.getSubject();
+        UserInfo userInfo = (UserInfo)currentUser.getPrincipal();
+        if (StringUtils.equals(userInfo.getRoles().get(0).getRole(), "admin")) {
+            User loginUser = userService.findByName(userInfo.getUsername());
+            Review review = new Review();
+            review.setModule(ReviewConstant.MODULE_USER);
+            review.setOperate(operate);
+            review.setOperateTime(new Date());
+            review.setOperatorId(loginUser.getId());
+            review.setOperatorName(loginUser.getRealName());
+            review.setReviewContent(JSON.toJSONString(user));
+            reviewService.addReview(review);
+            return true;
+        }
+        return false;
+    }
+
+    private boolean commitReviews(List<Long> ids, String operate) {
+        Subject currentUser = SecurityUtils.getSubject();
+        UserInfo userInfo = (UserInfo)currentUser.getPrincipal();
+        if (StringUtils.equals(userInfo.getRoles().get(0).getRole(), "admin")) {
+            User loginUser = userService.findByName(userInfo.getUsername());
+            List<User> users = userService.findByIds(ids);
+            users.forEach(user -> {
+                Review review = new Review();
+                review.setModule(ReviewConstant.MODULE_USER);
+                review.setOperate(operate);
+                review.setOperateTime(new Date());
+                review.setOperatorId(loginUser.getId());
+                review.setOperatorName(loginUser.getRealName());
+                review.setReviewContent(JSON.toJSONString(user));
+                reviewService.addReview(review);
+            });
+            return true;
+        }
+        return false;
     }
 }
